% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA: Principal Component Analysis function.}
\usage{
PCA(X)
}
\arguments{
\item{X}{High dimensional data}
}
\value{
Principal components matrix (output of eigen vectors)
}
\description{
This function performs principal component analysis, a dimensionality-reduction method that is often used to reduce the dimensionality of large data sets.
}
\examples{
PCA(X=zip.train)

PCA(X=zip.test)

}
